api-django/requirements.txt
arabic-reshaper==2.1.1
asgiref==3.3.1
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.2
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
Markdown==3.3.4
Pillow==8.1.0
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.1
reportlab==3.5.59
six==1.15.0
sqlparse==0.4.1
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
